Addressing maintainer comments
