+++ tests/instruments/test_all_instruments.py
@@ -140,7 +140,6 @@ grandfathered_docstring_instruments = [
     "SR830",
     "SR860",
     "SR510",
-    "TSL550",
 ]
